requirements.txt
aiohttp>=3.9.0
dnspython>=2.4.0
asyncpg>=0.29.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
